Add Chutney to the volunteer page
By contrast with ExitMap I'm a tad less sold on Chutney, but we'll see where it
goes. Nick put quite a bit of work into it in 2014, but not much since then. As
I understand it Chutney is due for an overhaul to move from 'tool for Nick' to
a more generally useful component.

Maybe I'm wrong.

If it gets more development attention then great, but if not I might drop it
back off the page later.