Drop volunteer page footer idea blob
This is an antique, horribly unreadable wad of text as denser than a legal
footer. We should invest effort in making project ideas readable or not have
them at all. Actually, this section predates me (meaning it's over six years
old) and never been maintained in that time. Guess it's so dense even we
ignored it. :P

If we want to resurrect any of these then it should be done so as tickets.