unlink those two obsolete pages
also, link the corepeople page as a sub-bullet, because that's
how it's done on the documentation pages. now there are two things
you can click that get you to corepeople. this is not perfect,
but it's better than the alternative where you don't even realize
it's a thing you can click on.